Fix point based occlusion artifacts ("surface acne")
For exact microrendering of close by disks, calculating quadratic form
coefficients with 32 bit floats leads to artifacts.  (In particular, some
disks appear to incorrectly self-shadow.)  Switching to using doubles
is easy and seems to fix the issue, at least for cases where the problem
has been observed so far.

+++ libs/pointrender/microbuffer.cpp
@@ -94,15 +94,15 @@ inline bool sphereOutsideCone(V3f p, float plen2, float r,
 ///
 /// The equation is assumed to have real-valued solutions; if they are in fact
 /// complex only the real parts are returned.
-inline void solveQuadratic(float a, float b, float c,
-                           float& lowRoot, float& highRoot)
+inline void solveQuadratic(double a, double b, double c,
+                           double& lowRoot, double& highRoot)
 {
     // Avoid NaNs when determinant is negative by clamping to zero.  This is
     // the right thing to do when det would otherwise be zero to within
     // numerical precision.
-    float det = std::max(0.0f, b*b - 4*a*c);
-    float firstTerm = -b/(2*a);
-    float secondTerm = sqrtf(det)/(2*a);
+    double det = std::max(0.0, b*b - 4*a*c);
+    double firstTerm = -b/(2*a);
+    double secondTerm = sqrtf(det)/(2*a);
     lowRoot = firstTerm - secondTerm;
     highRoot = firstTerm + secondTerm;
 }
@@ -144,7 +144,7 @@ static void renderDiskExact(IntegratorT& integrator, V3f p, V3f n, float r)
         //
         //   q(u,v) = a0*u*u + b0*u*v + c0*v*v + d0*u + e0*v + f0
         //
-        // such that the disk lies in the region satisfying f(u,v) < 0.  Start
+        // such that the disk lies in the region satisfying q(u,v) < 0.  Start
         // with the implicit definition of the disk on the plane,
         //
         //   norm(dot(p,n)/dot(V,n) * V - p)^2 - r^2 < 0
@@ -153,45 +153,53 @@ static void renderDiskExact(IntegratorT& integrator, V3f p, V3f n, float r)
         //
         //   A*dot(V,V) + B*dot(V,n)*dot(p,V) + C < 0
         //
-        float dot_pn = dot(p,n);
-        float A = dot_pn*dot_pn;
-        float B = -2*dot_pn;
-        float C = plen2 - r*r;
+        // NOTE the use of double precision.  This is used intentionally to
+        // avoid "surface acne" artifacts (looks much like classic shadow map
+        // self-shadowing).  The underlying cause is probably cancellation
+        // leading to catastrophic loss of precision, but it's much easier to
+        // patch the problem with doubles than analyze exactly where in the
+        // following mess of coefficients the problem ultimately happens!  Note
+        // that deferring to the simple raytracing code below also fixes the
+        // issue.
+        double dot_pn = dot(p,n);
+        double A = dot_pn*dot_pn;
+        double B = -2*dot_pn;
+        double C = plen2 - r*r;
         // Next, project this onto the current face to compute the
-        // coefficients a0 through to f0.
+        // coefficients a0 through to f0 for q(u,v)
         V3f pp = MicroBuf::canonicalFaceCoords(iface, p);
         V3f nn = MicroBuf::canonicalFaceCoords(iface, n);
-        float a0 = A + B*nn.x*pp.x + C*nn.x*nn.x;
-        float b0 = B*(nn.x*pp.y + nn.y*pp.x) + 2*C*nn.x*nn.y;
-        float c0 = A + B*nn.y*pp.y + C*nn.y*nn.y;
-        float d0 = (B*(nn.x*pp.z + nn.z*pp.x) + 2*C*nn.x*nn.z);
-        float e0 = (B*(nn.y*pp.z + nn.z*pp.y) + 2*C*nn.y*nn.z);
-        float f0 = (A + B*nn.z*pp.z + C*nn.z*nn.z);
+        double a0 = A + B*nn.x*pp.x + C*nn.x*nn.x;
+        double b0 = B*(nn.x*pp.y + nn.y*pp.x) + 2*C*nn.x*nn.y;
+        double c0 = A + B*nn.y*pp.y + C*nn.y*nn.y;
+        double d0 = (B*(nn.x*pp.z + nn.z*pp.x) + 2*C*nn.x*nn.z);
+        double e0 = (B*(nn.y*pp.z + nn.z*pp.y) + 2*C*nn.y*nn.z);
+        double f0 = (A + B*nn.z*pp.z + C*nn.z*nn.z);
         // Finally, transform the coefficients so that they define the
         // implicit function in raster face coordinates, (iu, iv)
-        float scale = 2.0f/faceRes;
-        float scale2 = scale*scale;
-        float off = 0.5f*scale - 1.0f;
-        float a = scale2*a0;
-        float b = scale2*b0;
-        float c = scale2*c0;
-        float d = ((2*a0 + b0)*off + d0)*scale;
-        float e = ((2*c0 + b0)*off + e0)*scale;
-        float f = (a0 + b0 + c0)*off*off + (d0 + e0)*off + f0;
+        double scale = 2.0f/faceRes;
+        double scale2 = scale*scale;
+        double off = 0.5f*scale - 1.0f;
+        double a = scale2*a0;
+        double b = scale2*b0;
+        double c = scale2*c0;
+        double d = ((2*a0 + b0)*off + d0)*scale;
+        double e = ((2*c0 + b0)*off + e0)*scale;
+        double f = (a0 + b0 + c0)*off*off + (d0 + e0)*off + f0;
         // The coefficients a,b,c,d,e,f may represent an ellipse or a
         // hyperbola in the face's raster coordinates, determine which.
-        float det = 4*a*c - b*b;
+        double det = 4*a*c - b*b;
         if(det > 0)
         {
             // If the coefficients represent an ellipse, we may construct a
             // tight bound.
             int ubegin = 0, uend = faceRes;
             int vbegin = 0, vend = faceRes;
-            float ub = 0, ue = 0;
+            double ub = 0, ue = 0;
             solveQuadratic(det, 4*d*c - 2*b*e, 4*c*f - e*e, ub, ue);
             ubegin = std::max(0, Imath::ceil(ub));
             uend   = std::min(faceRes, Imath::ceil(ue));
-            float vb = 0, ve = 0;
+            double vb = 0, ve = 0;
             solveQuadratic(det, 4*a*e - 2*b*d, 4*a*f - d*d, vb, ve);
             vbegin = std::max(0, Imath::ceil(vb));
             vend   = std::min(faceRes, Imath::ceil(ve));
@@ -202,7 +210,7 @@ static void renderDiskExact(IntegratorT& integrator, V3f p, V3f n, float r)
             for(int iv = vbegin; iv < vend; ++iv)
             for(int iu = ubegin; iu < uend; ++iu)
             {
-                float q = a*(iu*iu) + b*(iu*iv) + c*(iv*iv) + d*iu + e*iv + f;
+                double q = a*(iu*iu) + b*(iu*iv) + c*(iv*iv) + d*iu + e*iv + f;
                 if(q < 0)
                 {
                     V3f d = integrator.rayDirection(iface, iu, iv);
